internal: Define the internal request data type
This commit defines the struct yhttp_requ_internal data type, which
is going to be used for the internal member field of struct yhttp_requ.

The yhttp_requ_init() and yhttp_requ_free() alongside with their
accompanying regression test have been modified accordingly, in order
to respect this field now.