parser: Begin work on parser_parse()
This commit adds the prototype and a dummy declaration for the core
routine of the HTTP request parser: parser_parse().

This specific function is called by the low-level TCP networking
functions every time a new TCP message has been received.
The parser_parse() function will then pass the new data to either
parse_rline(), parser_headers() or parser_body(), based on the value of
parser->state.

Once parser_parse() has been called, it will immediately begin to hand
the control over to a function responsible for the parsing of the
current part of the HTTP request (parser_rline(), parser_headers(),
parser_body()).

These functions will then look for the end of that specific part of the
HTTP request, meaning parser_rline() will look for CRLF,
parser_headers() for CRLFCRLF, and so on.

In case that the end has not been found inside the just received TCP
message, it is going to be appended to an internal buffer.

Otherwise, it is also going to be added to the buffer, which is then
going to be parsed up until that specific end.  The function will
then wipe the buffer up until the end and set parser->state to the
next state of the HTTP request parsing process.