buf: Implement buf_copy()
This commit implements the buf_copy() function, which creates a deep
copy of one buffer.

The functionality for this is going to be required for the HTTP
request parser, because the received TCP messages need to splitted
sooner or later.  More of this is going to be explained within
future code comments and commit messages.