internal: Change the way how yhttp_stop() works
This commit removes the quit field from struct yhttp and replaces it
with a pipe(2).

The motivation behind this is, that yhttp_stop() will only work if the
poll(2) system call has been interrupted by a signal.  This results in
the fact, that yhttp_stop() can only be called from within a signal
handler and not by other "concurrent" mechanisms such as threads.

In order to address this issue, yhttp_dispatch() will now create a
pipe(2), which yhttp_stop() will use to deliver an EOF then.