net: Access the sockets through struct poll_data
This commit modifies the net_handle_accept() and net_handle_client()
functions to accept the index of the appropriate socket inside
poll_data.pfds rather than the plain socket file descriptor.

The motivation behind this is, that the struct poll_data will
eventually contain other data, such as struct parser, using the same
indices as in the pfds array.