net: Add parsers array to struct poll_data
This commit adds an array holding struct parser * to the
struct poll_data.

The array shares its size (allocated space) and indices with the pfds
array, meaning that the index of a client inside pfds also refers to
its parser instance inside parsers.

It is necessary to have this, because each connected client needs its
own instance of a parser.  It is only logical storing it alongside with
pfds, as this already serves as a way to identify the client.