net: Optimize net_poll_del()
This commit optimizes the net_poll_del() function, by changing its
prototype to now accept its index inside the struct poll_data array
instead of the file descriptor.

Because of this, a loop performing a linear search is being removed,
as the index of the file descriptor inside the struct poll_data array
is already known and does not need to be looked-up first.