net: Implement net_poll_close()
This commit implements the net_poll_close() function, which removes a
file descriptor from the struct poll_data array and closes it
afterwards.  Effectively, it is only a chain of a net_poll_del() and
a close(2) call.

The reason for this is, that the code base currently contains many
manual calls to net_poll_del() and a subsequent close(2).  This alone
is a reason for putting this into its own function.