parser: Implement parser_find_eol()
This commit implements the parser_find_eol() function, which looks for
the first character terminating a line, which is either LF or, if the
next character is LF, CR.

It will be required for various parsers inside the HTTP header, most
notably the rline and the header fields.

Beside this, this commit also implements an accompanying regression
test.