parser: Implement the HTTP target parser
This commit extends the HTTP parser, by implementing the functions
relevant for parsing the target of the rline.

Because the target consists of two components, namely the path and the
query, the complexity was splitted among two functions:
parser_rline_path() and parser_rline_query(), which just calls the more
complex parser_query(), in which the real action happens.

The inner workings of parser_rline_path() are quite simple, so no
further elaboration is being given on that.

The lions' share of complexity however, is within the parser_query()
function, which extracts the various key/value pairs of a query string
into the queries hash table from the internal requests field.

Currently, there is no regression test for any of these functions.
This is a serious problem, which is hopefully going to be addressed
within the next commits.