parser: Implement parser_rline()
This commit implements the parser_rline() function, which performs a
rudimentary parsing of the rline by searching for the spaces.  After
this, it calls parser_rline_method() and parser_rline_target() with the
lengths and starting points obtained in the previous step.  Afterwards,
it cuts the parsed part in the buffer away and sets the state to
PARSER_HEADERS.  The HTTP-version is currently ignored, because it has
no relevance to us yet.

Beside this, this commit also implements an accompanying regression
test.