parser: Implement parser_body()
This commit implements the dead-simple parser_body() function, which
does nothing but changes the state to PARSER_DONE and sets the
parser->requ->body pointer to the parser->buf.buf address, once the
buffer length reaches the value of the supplied Content-Length header
field.

Because of this design decision, the yhttp_requ_free() function had to
be modified slightly to no longer free(3) the requ->body pointer, as its
ownership is no longer in its scope.