internal: Define the internal response data type
This commit defines the struct yhttp_resp data type, which makes up a
response to a request and is thereby also included in the struct
yhttp_requ_internal.

This data type is being initialized with the yhttp_resp_init() and
yhttp_resp_free() functions, both of which are called inside
yhttp_requ_init() and yhttp_requ_free() respectively.

Beside this, this commit also adds an accompanying regression test and
modifies the one for yhttp_requ_init() and yhttp_requ_free()
accordingly.