hash: Be case insensitive
This commit adjusts the hash() and hash_get() function to ignore the
case for the name.

The reason for this is, that the hash table is required for data
structures, where the case does not matter anyway, such as for query
strings or header fields.

Previously, header field names have been converted to lower-case which
resulted in all header fields being outputted with their lower-case
names which was bad.  This is no longer the case (no pun!) with this
commit.