parser: Remove PARSER_ERR
This commit removes the PARSE_ERR state from the enum parser_state, as
announced in f758c58 and replaces it with the err member field of struct
parser alone.

The reason for this is, that when the parser was initially designed, the
author did not considered the option for custom error codes, which is by
far a much better design choice.  Thereby, the code base for error
detection during the parsing was based on the state, rather than the
presence of an error code.

Now, the checks against PARSER_ERR have all been replaced by checks
against parser->err (treating it as a boolean, as zero is its default
value).