util: Implement util_aprintf()
This commit implements the util_aprintf() function, a function similar
to asprintf(3) found in glibc and OpenBSDs libc, with the notable
exception of directly returning the allocated string rather than an
integer.

It is going to be necessary in the response generation, as various
things have to be generated in a formatted way, for which the
printf(3)-functions offer a great solution.

Beside this, this commit also implements an accompanying regression
test.