net: Implement net_send()
This commit implements the net_send() function, a function for sending
large amounts of binary data.

The reason for this is, that send(2) cannot be used directly, as it does
not guarantee to actually send the data it has been instructed to, so
there is a need for a little helper functions, which invokes send(2)
until all data has been send, and this is where net_send() comes into
use.