resp: Implement resp()
This commit implements the resp() function, a function for formatting
and transmitting the user generated response to a request.

For this, the resp() function makes use of various helper functions,
most notably the resp_transmit_* functions, that format the response by
making use of the resp_fmt_* functions, and transmit them using the
net_send() function.