resp: Implement resp_err()
This commit implements the resp_err() function, a function for
formatting and transmitting an error message based on a status code.

It does so, by utilizing the accompanying resp_fmt_err() function, which
generates an rline with the status code and reason-phrase as well as a
message body with the exact same reason phrase and a header with a
matching Content-Length.

Beside this, the regress/test-resp_fmt regression test has been updated
accordingly.