parser: Reimplement the header parser from scratch
This commit reimplements the header parser from scratch, this time with
support for OWS.

It became clear, that the header parser required a new implementation,
as the old one was very messy, did not supported OWS and was even
accessing uninitialized content.  This is an approach to address those
issues.