TODO: Explain bug in parser_headers()
Currently there is a critical bug within the parser_headers() function,
which makes receiving the header in multiple TCP frames impossible.  The
reason for this bug is, that the parser_headers() function sets the sol
variable to the start of the buffer, no matter how many lines have
actually been parsed yet, meaning that the header parser always start
from scratch at every new TCP frame, resulting in a malformatted
request, because header fields are being than set to be duplicates.

A solution to this issue would either be the usage of buf_pop() or by
storing sol or something similar directly in struct parser.