parser: Fix segmentation bug in parser_headers()
This commit fixes the TCP segmentation bug explained in 828e344, by
modifying the way how parser_headers() works.

Previously, parser_headers() went through all received lines found in
the buffer, until either the end of the header or a not-fully received
line has been found.  The latter was the condition for the bug, as
further calls to parser_headers() all began parsing at the start of the
buffer again, resulting in the detection of duplicate field-names.

Now, the parser_headers() function parses if and only if the entire
HTTP header has been received.  Otherwise, it will just return YHTTP_OK
and wait for more input.  This approach has not been explained in the
828e344 commit as a possible solution, but turns out to be the most
elegant one, as storing sol in struct parser would be a weird exception
and various calls to buf_pop() would result in a terrible performance.