net: Replace net_ip4() and net_ip6() by net_ip()
This commit removes the net_ip4() and net_ip6() functions in favor for
the new net_ip() function, which is agnostic to the socket domain.

The reason for this is, that this reduces the amount of code and storage
required to determine the socket domain.  Beside this, it also reduces
the amount of duplicate code and is generally a more cleaner way.