net: Return with YHTTP_OK if the response failed
This commit fixes a critical vulnerability introduced by 794dc0d, namely
an exploit that would have allowed an attacker to terminate the
execution of the server.

The reason for this was, that the net_handle_client() function would
have returned with the error code of resp() and resp_err() and not
YHTTP_OK, resulting in net_dispatch() and yhttp_dispatch() to fail with
that error code as well.  Because these two functions return an error
if the transmission has failed, an attacker could have simply triggered
such an error by terminating her connection during a transfer.