net: Check for EAGAIN in net_send()
This commit adds a check for EAGAIN in net_send() in order to try again,
if it occurred.

Previously, it was impossible to send large amounts of data through net_send(),
because the function failed with EAGAIN, which was then interpreted as a fatal
error.  This is completely unacceptable behavior in an application using
non-blocking TCP sockets.